sentry - routes - new `routers` module: - analytics_router - move the module - analytics - fix tests and move/use analtyics_router instead of analytics
