+++ sentry/src/routes.rs
@@ -224,3 +224,5 @@ pub mod campaign;
 // `cfg` module has single request, so we only export this request
 mod cfg;
 pub mod channel;
+
+pub mod routers;
